Update README.md
The discrete two-dimensional fast wavelet transform was performed on the hyperspectral image to enhance the spectral information as a spectral feature for classifying tree species. In parallel, the horizontal, vertical, and diagonal details of the high-frequency sub-bands were extracted to obtain the texture characteristics of the three directions of the hyperspectral image.